add max values for parameters

+++ Generator/ImageGenerator/imagegenerator.cpp
@@ -12,6 +12,6 @@ using namespace terraingenerator;
 
 void terraingenerator::export_png(const Terrain& terrain, const char *file_name)
 {
-    stbi_write_png(file_name, (int) terrain.height_, (int) terrain.width_,
+    stbi_write_png(file_name, terrain.height_, terrain.width_,
                    1, terrain.heightmap_, terrain.height_);
 }
